+++ board/stm32h7/lladc.h
@@ -20,7 +20,7 @@ void adc_init(void) {
 uint16_t adc_get_raw(uint8_t channel) {
 
   ADC1->SQR1 &= ~(ADC_SQR1_L);
-  ADC1->SQR1 = (channel << 6U);
+  ADC1->SQR1 = ((uint32_t) channel << 6U);
 
   ADC1->SMPR1 = (0x7U << (channel * 3U) );
   ADC1->PCSEL_RES0 = (0x1U << channel);
